check remove which are not required

+++ src/main/java/com/main/glory/servicesImpl/StockBatchServiceImpl.java
@@ -242,6 +242,9 @@ public class StockBatchServiceImpl {
         if(Objects.equals(stockMast.get().getIsProductionPlanned(),true)){
             throw new Exception("Can't delete the batch, already in production, for id:"+id);
         }
+        List<ProductionPlan> productionPlans = productionPlanService.getProductionByStockId(id);
+        if(!productionPlans.isEmpty())
+            throw new Exception("can't delete the stock , because already sent to production");
 
         stockMastDao.deleteById(id);
     }
